Tile Size auf double-Werte geändert, sodass alle Bildschirmgrößen aktuell 48 Blöcke darstellen. (Bitte klappt es)

+++ game/objects/GameObject.java
@@ -3,6 +3,6 @@ package game.objects;
 import java.awt.*;
 
 public abstract class GameObject {
-
-	public abstract void render(Graphics2D g, int tileSize);
+	//Methoden
+	public abstract void render(Graphics2D g, double tileSize);
 }
