+++ game/objects/creatures/Player.java
@@ -12,7 +12,7 @@ public class Player extends Creature{
 	}
 	
 	@Override
-	public void render(Graphics2D g, int tileSize) {
+	public void render(Graphics2D g, double tileSize) {
 		double centerXOnScreen = centerX * tileSize;
 		double centerYOnScreen = centerY * tileSize;
 		double diameterOnScreen = radius * 2.0;
